requirements.txt
openai>=1.0.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
python-multipart>=0.0.6
requests>=2.31.0
PyGithub>=2.1.0